requirements.txt
asgiref==3.8.1
autopep8==2.3.1
colorama==0.4.6
Django==5.0.6
django-silk==5.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gprof2dot==2024.6.6
haversine==2.8.1
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.0
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
